# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "wi-Q Great Food Integration",
    "summary": "Consumes Great Food Ltd menu APIs and exposes scenario workflows.",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Health Check",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Check Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/auth_token": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Create Auth Token",
        "operationId": "create_auth_token_auth_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_auth_token_auth_token_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessToken"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/menus": {
      "get": {
        "tags": [
          "menus"
        ],
        "summary": "List Menus",
        "operationId": "list_menus_menus_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_list_Menu__"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/menu/{menu_id}/products": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "List Products For Menu",
        "operationId": "list_products_for_menu_menu__menu_id__products_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "menu_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Menu Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_list_Product__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/menu/{menu_id}/product/{product_id}": {
      "put": {
        "tags": [
          "products"
        ],
        "summary": "Rename Product",
        "operationId": "rename_product_menu__menu_id__product__product_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "menu_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Menu Id"
            }
          },
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Product Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductRenameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_Product_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AccessToken": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type"
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_in",
          "token_type",
          "scope"
        ],
        "title": "AccessToken"
      },
      "Body_create_auth_token_auth_token_post": {
        "properties": {
          "client_secret": {
            "type": "string",
            "title": "Client Secret"
          },
          "client_id": {
            "type": "string",
            "title": "Client Id"
          },
          "grant_type": {
            "type": "string",
            "title": "Grant Type"
          }
        },
        "type": "object",
        "required": [
          "client_secret",
          "client_id",
          "grant_type"
        ],
        "title": "Body_create_auth_token_auth_token_post"
      },
      "DataResponse_Product_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Product"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "DataResponse[Product]"
      },
      "DataResponse_list_Menu__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Menu"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "DataResponse[list[Menu]]"
      },
      "DataResponse_list_Product__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Product"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "DataResponse[list[Product]]"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Menu": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "Menu"
      },
      "Product": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "Product"
      },
      "ProductRenameRequest": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ProductRenameRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}